requirements.txt
# Core Dependencies
python>=3.11

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# LangChain
# 최신 버전 사용 (0.1.0+)
langchain>=0.1.0
langchain-community>=0.0.10
# 특정 버전 고정이 필요한 경우 아래 주석 해제
# langchain==0.1.12
# langchain-community==0.0.20

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# OpenAI (optional)
openai>=1.3.0

# Document Parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0  # Alternative PDF parser

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Configuration
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0

